Auth working on api routes

+++ src/app/api/articles/route.ts
@@ -1,21 +1,29 @@
 import connectMongoDB from "@/libs/mongodb";
 import Article from "@/models/article";
+import { getServerSession } from "next-auth";
+import { getToken } from "next-auth/jwt";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(request: NextRequest){
+    const session = await getServerSession()
+    if(!session) return NextResponse.json({message: "Not authenticated"}, {status: 401})
+        
     const {title, perex, date} = await request.json();
     await connectMongoDB();
     await Article.create({title, perex, date});
     return NextResponse.json({message: "Article Created"}, {status: 201});
 }
 
-export async function GET(){
+export async function GET(req: NextRequest){
     await connectMongoDB();
     const articles = await Article.find();
     return NextResponse.json({articles});
 }
 
 export async function DELETE(request: NextRequest) {
+    const session = await getServerSession()
+    if(!session) return NextResponse.json({message: "Not authenticated"}, {status: 401})
+
     const id = request.nextUrl.searchParams.get("id");
     await connectMongoDB();
     await Article.findByIdAndDelete(id);
